Zadanie: testy jednostkowe dla kalkulatora

+++ kodilla-basic-test/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -9,8 +9,7 @@ public class Calculator {
     public int subtract(int a, int b) {
         return a - b;
     }
-    public int power(int a){
+    public double power(double a){
         return a*a;
     }
 }
-
